feed-master

+++ voyageone-static/2.0/develop/modules/cms/views/channel/imagegrouplist/index.ctl.js
@@ -3,21 +3,26 @@
  */
 define([
     'cms',
+    'underscore',
     'modules/cms/controller/popup.ctl'
 ], function (cms) {
     cms.controller("imageGroupController", (function () {
         function ImageGroupController($routeParams, imageGroupService, cActions, confirm, alert, notify) {
+            this.confirm = confirm;
+            this.alert = alert;
+            this.notify = notify;
+
             this.platformList = [];
             this.imageType = "";
             this.beginModified = "";
             this.endModified = "";
             this.brandNameList = [];
 
-            this.brandName = "";
+            this.brandName = [];
             this.productTypeList = [];
-            this.productType = "";
+            this.productType = [];
             this.sizeTypeList = [];
-            this.sizeType = "";
+            this.sizeType = [];
 
             this.imageGroupList = [];
             this.pageOption = {
@@ -53,17 +58,35 @@ define([
                     "productType" : main.productType,
                     "sizeType" : main.sizeType
                 }).then(function (res) {
-                        main.imageGroupList = res.data.imageGroupList;
-                        main.pageOption.total = res.data.imageGroupList.length;
+                        main.imageGroupList = res.data
+                        main.pageOption.total = res.data.length;
                         main.pageOption.curr = 1;
                 })
             },
+            clear: function () {
+                this.imageType = "";
+                this.beginModified = "";
+                this.endModified = "";
+                this.brandName = [];
+                this.productType = [];
+                this.sizeType = [];
+                _.each(this.platformList, function (platform) {
+                    platform.show = false;
+                });
+            },
             delete: function (imageGroupId) {
                 var main = this;
-                main.imageGroupService.delete({
-                    "imageGroupId" : imageGroupId
-                }).then(function (res) {
-
+                main.confirm('TXT_MSG_DO_DELETE').result.then(function () {
+                    main.imageGroupService.delete({
+                        "imageGroupId" : imageGroupId
+                    }).then(function (res) {
+                        main.notify.success('TXT_MSG_DELETE_SUCCESS');
+                        main.search();
+                    }, function (err) {
+                        if (err.displayType == null) {
+                            main.alert('TXT_MSG_DELETE_FAIL');
+                        }
+                    })
                 })
             }
         };
